requirements.txt
langchain
streamlit
fitz
langchain-openai
python-dotenv
faiss-cpu